Separate UI and logic of pinCode
